gistda_to_gcs/requirements.txt
google-cloud-storage
requests
python-dotenv